requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.0
h11==0.14.0
httptools==0.6.4
idna==3.10
./vendor/payos-0.1.8.tar.gz
pydantic==2.11.2
pydantic_core==2.33.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.38.6
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
